% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-functions.R
\name{geo_mean}
\alias{geo_mean}
\title{Geometric mean}
\usage{
geo_mean(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector.}

\item{na.rm}{Whether to remove missing values.}
}
\value{
Geometric mean of positive values.
}
\description{
Geometric mean
}
