% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-functions.R
\name{\%ni\%}
\alias{\%ni\%}
\title{Negated value matching}
\usage{
x \%ni\% table
}
\arguments{
\item{x}{Values to match.}

\item{table}{Values to match against.}
}
\value{
Logical vector indicating values not found in \code{table}.
}
\description{
Negated value matching
}
