% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize a boosterpak project}
\usage{
init(
  root = ".",
  renv = c("ask", "yes", "no"),
  rprofile = c("ask", "yes", "no"),
  verbose = NULL
)
}
\arguments{
\item{root}{Project root.}

\item{renv}{Whether to initialize project-local renv: \code{"ask"}, \code{"yes"}, or \code{"no"}.}

\item{rprofile}{Whether to add the helper auto-source line to \code{.Rprofile}.}

\item{verbose}{Whether to print routine summaries.}
}
\value{
Project setup paths, invisibly.
}
\description{
Initialize a boosterpak project
}
