% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_functions}
\alias{list_functions}
\title{List available materializable functions}
\usage{
list_functions(root = ".", verbose = NULL)
}
\arguments{
\item{root}{Project root.}

\item{verbose}{Whether to print a routine summary.}
}
\value{
A data frame of available functions and installation status, invisibly.
}
\description{
List available materializable functions
}
