% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packs.R
\name{list_packs}
\alias{list_packs}
\title{List available packs}
\usage{
list_packs(scope = NULL, root = ".", verbose = NULL)
}
\arguments{
\item{scope}{Optional scope filter: \code{"project"}, \code{"user"}, or \code{"builtin"}.}

\item{root}{Project root.}

\item{verbose}{Whether to print a routine summary.}
}
\value{
A data frame of available packs, invisibly.
}
\description{
List available packs
}
