% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-functions.R
\name{my_skim}
\alias{my_skim}
\title{Skimr preset with numeric helpers}
\usage{
my_skim(data, ...)
}
\arguments{
\item{data}{Data frame to skim.}

\item{...}{Additional arguments passed to \code{skimr::skim()}.}
}
\value{
A skimr summary.
}
\description{
Skimr preset with numeric helpers
}
