\name{promote_pack}
\alias{promote_pack}
\title{Promote a project pack to user scope}
\usage{
promote_pack(name, root = ".", overwrite = FALSE, verbose = NULL)
}
\arguments{
\item{name}{Pack name.}

\item{root}{Project root.}

\item{overwrite}{Whether to replace an existing user-scope pack.}

\item{verbose}{Whether to print routine summaries.}
}
\value{
Path to the copied user-scope pack, invisibly.
}
\description{
Promote a project pack to user scope
}
