% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{status}
\alias{status}
\title{Report boosterpak project status}
\usage{
status(root = ".", verbose = NULL)
}
\arguments{
\item{root}{Project root.}

\item{verbose}{Whether to print routine summaries.}
}
\value{
A list describing project status, invisibly.
}
\description{
Report boosterpak project status
}
