% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-functions.R
\name{theme_custom}
\alias{theme_custom}
\title{Custom ggplot2 theme}
\usage{
theme_custom(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A ggplot2 theme.
}
\description{
Custom ggplot2 theme
}
